split implementation for shopping cart, change package name
