+++ internal/shopping_cart/handler/shopping_cart_handler.go
@@ -1,77 +1,49 @@
-package shopping_cart
+package handler
 
 import (
 	model "Dzaakk/simple-commerce/internal/shopping_cart/models"
 	usecase "Dzaakk/simple-commerce/internal/shopping_cart/usecase"
-	auth "Dzaakk/simple-commerce/package/auth"
 	template "Dzaakk/simple-commerce/package/template"
 	"fmt"
 	"net/http"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
-	"github.com/go-redis/redis/v8"
 )
 
-type ShoppingCarthandler struct {
+type ShoppingCartHandler struct {
 	Usecase usecase.ShoppingCartUseCase
 }
 
-func NewShoppingCartHandler(usecase usecase.ShoppingCartUseCase) *ShoppingCarthandler {
-	return &ShoppingCarthandler{
+func NewShoppingCartHandler(usecase usecase.ShoppingCartUseCase) *ShoppingCartHandler {
+	return &ShoppingCartHandler{
 		Usecase: usecase,
 	}
 }
-func (handler *ShoppingCarthandler) Route(r *gin.RouterGroup, redis *redis.Client) {
-	ShoppingHandler := r.Group("api/v1")
 
-	ShoppingHandler.Use()
-	{
-		ShoppingHandler.POST("/shopping-cart", auth.JWTMiddleware(redis), func(ctx *gin.Context) {
-			if err := handler.AddProductToShoppingCart(ctx); err != nil {
-				ctx.JSON(http.StatusInternalServerError, template.Response(http.StatusInternalServerError, "internal server error", err.Error()))
-				return
-			}
-		})
-		ShoppingHandler.GET("/shopping-cart", auth.JWTMiddleware(redis), func(ctx *gin.Context) {
-			if err := handler.GetListShoppingCart(ctx); err != nil {
-				ctx.JSON(http.StatusInternalServerError, template.Response(http.StatusInternalServerError, "failed to Get product", err.Error()))
-				return
-			}
-		})
-		ShoppingHandler.POST("/shopping-cart/delete", auth.JWTMiddleware(redis), func(ctx *gin.Context) {
-			if err := handler.DeleteShoppingList(ctx); err != nil {
-				ctx.JSON(http.StatusInternalServerError, template.Response(http.StatusInternalServerError, "failed to Get product", err.Error()))
-				return
-			}
-		})
-	}
-}
-
-func (handler *ShoppingCarthandler) AddProductToShoppingCart(ctx *gin.Context) error {
+func (handler *ShoppingCartHandler) AddProductToShoppingCart(ctx *gin.Context) {
 	var reqData model.ShoppingCartReq
 	if err := ctx.ShouldBindJSON(&reqData); err != nil {
 		ctx.JSON(http.StatusBadRequest, template.Response(http.StatusBadRequest, "Bad Request", "Invalid input data"))
-		return nil
+		return
 	}
 
 	template.AuthorizedChecker(ctx, reqData.CustomerId)
 	newShopingCart, err := handler.Usecase.Add(reqData)
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, template.Response(http.StatusInternalServerError, "Internal Server Error", err.Error()))
-		return nil
+		return
 	}
 
 	ctx.JSON(http.StatusCreated, template.Response(http.StatusOK, "Success", newShopingCart))
-	return nil
 }
 
-func (handler *ShoppingCarthandler) GetListShoppingCart(ctx *gin.Context) error {
+func (handler *ShoppingCartHandler) GetListShoppingCart(ctx *gin.Context) {
 	customerId, _ := strconv.Atoi(ctx.Query("customerId"))
 
 	template.AuthorizedChecker(ctx, ctx.Query("customerId"))
 	if ctx.IsAborted() {
-		return nil
+		return
 	}
 	listShoppingCart, err := handler.Usecase.GetListItem(customerId)
 	if err != nil {
@@ -87,31 +59,29 @@ func (handler *ShoppingCarthandler) GetListShoppingCart(ctx *gin.Context) error 
 		}
 
 		ctx.JSON(statusCode, template.Response(statusCode, message, err.Error()))
-		return nil
+		return
 	}
 
 	ctx.JSON(http.StatusCreated, template.Response(http.StatusOK, "Success", listShoppingCart))
-	return nil
 }
 
-func (handler *ShoppingCarthandler) DeleteShoppingList(ctx *gin.Context) error {
+func (handler *ShoppingCartHandler) DeleteShoppingList(ctx *gin.Context) {
 	var data model.DeleteReq
 	if err := ctx.ShouldBindJSON(&data); err != nil {
 		ctx.JSON(http.StatusBadRequest, template.Response(http.StatusBadRequest, "Bad Request", "Invalid input data"))
-		return nil
+		return
 	}
 	fmt.Printf("input = %v", data)
 	template.AuthorizedChecker(ctx, data.CustomerId)
 	if ctx.IsAborted() {
-		return nil
+		return
 	}
 
 	err := handler.Usecase.DeleteShoppingList(data)
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, template.Response(http.StatusInternalServerError, "Internal Server Error", err.Error()))
-		return nil
+		return
 	}
 
 	ctx.JSON(http.StatusCreated, template.Response(http.StatusOK, "Success", "Success Delete Shopping List"))
-	return nil
 }
