+++ internal/shopping_cart/models/shopping_cart.go
@@ -1,4 +1,4 @@
-package shopping_cart
+package models
 
 import "Dzaakk/simple-commerce/package/template"
